ui: change search icon to FAB
