+++ app/src/main/java/com/nelsonxilv/gstoutimetable/presentation/screen/TimetableUiState.kt
@@ -14,6 +14,6 @@ sealed interface TimetableUiState {
 
     data object Error : TimetableUiState
     data object Loading : TimetableUiState
-    data object EmptyTimetable: TimetableUiState
+    data object EmptyTimetable : TimetableUiState
     data object Hello : TimetableUiState
 }
